requirements.txt
setuptools==69.1.0
pip==24.0
asgiref==3.7.2
Django==5.0.6
sqlparse==0.4.4
tzdata==2024.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
requests==2.31.0
boto3==1.34.46
djangorestframework==3.14.0
gunicorn==21.2.0
humanfriendly==10.0
ipaddress==1.0.23
sentry-sdk==1.40.5
django-filter==24.2
django-user-agents==0.4.0
django-model-utils==4.5.1
shortuuid==1.0.11
phonenumbers==8.13.31
Pillow==10.3.0
cryptography==42.0.5
dropbox==11.36.2
pcloud==1.3
ibm-cos-sdk==2.13.4
slack-sdk==3.18.3
google-auth-oauthlib==1.2.0
azure-storage-blob==12.16.0
azure-identity==1.12.0
oss2==2.17.0
cos-python-sdk-v5==1.9.23
google-cloud-storage==2.17.0
humanize==4.3.0
oci==2.129.0
paramiko==3.0.0
ftputil==5.0.4
mysql-connector-python==8.0.30
beautifulsoup4==4.12.3
celery==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
gcloud==0.18.3
git+https://github.com/hellysmile/fake-useragent.git@460686d0c729c2b22172dc75760408c87b7d6618
git+https://github.com/ovh/python-ovh.git